Ahora los métodos se encargan por sí solos de acceder a la base de datos y cerrar la conexión con ella (los que lo necesiten).
